Split the tests. Rename remote scheduler files
